Default 4.5+ star filter, float stars, operator dropdown fix
- Default min_stars=4.5 when no star filter selected in URL
- Change hotel_stars column to REAL for4.5 support
- Store float stars in RedTag scraper
- Operator dropdown populated from all known tour operators
- Any stars option uses value=0 to distinguish from default

+++ models.py
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS deals (
     destination TEXT NOT NULL,
     deal_type TEXT DEFAULT 'hotel',
     hotel_name TEXT,
-    hotel_stars INTEGER,
+    hotel_stars REAL,
     departure_date TEXT,
     return_date TEXT,
     nights INTEGER,
